Чан: Клиент: Рефакторинг
