+++ frontend/README.md
@@ -1,4 +1,4 @@
-# vue-front-chan
+# schizochan-front
 
-* `npm i && npm run dev-serve` for dev
+* `npm i && npm run serve` for dev
 * `npm i && npm run build` for build prod files
