+++ frontend/src/components/Tracks.vue
@@ -47,6 +47,7 @@
 <script>
 import axios from 'axios'
 import { bus } from '../bus'
+import { formatDuration } from '../utils/duration_formatter'
 
 const config = require('../../config')
 const _      = require('lodash')
@@ -73,14 +74,7 @@ export default {
   },
   methods: {
     formatDuration: function (value) {
-      var sec_num = parseInt(value, 10);
-      var hours   = Math.floor(sec_num / 3600);
-      var minutes = Math.floor((sec_num - (hours * 3600)) / 60);
-      var seconds = sec_num - (hours * 3600) - (minutes * 60);
-      if (hours   < 10) {hours   = "0"+hours;}
-      if (minutes < 10) {minutes = "0"+minutes;}
-      if (seconds < 10) {seconds = "0"+seconds;}
-      return hours+':'+minutes+':'+seconds;
+      return formatDuration(value)
     },
     resetFilters: function (event) {
       event.preventDefault()
